[Refactor] 코드리뷰 반영 Rename - 이름뒤 UseCase 추가
