+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/DeleteFoodCountUseCase.kt
@@ -1,5 +1,5 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-interface EditFoodCount {
+interface DeleteFoodCountUseCase {
     operator fun invoke()
 }
