+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/DeleteFoodCountUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-class DeleteFoodCountImpl:DeleteFoodCount {
+class DeleteFoodCountUseCaseImpl:DeleteFoodCountUseCase {
     override fun invoke() {
         TODO("Not yet implemented")
     }
