+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetRecipeListUseCase.kt
@@ -1,5 +1,5 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-interface GetRecipeList { // 레피시이름 리스트를 가져오는 함수
+interface GetRecipeListUseCase { // 레피시이름 리스트를 가져오는 함수
     operator fun invoke():List<String>
 }
