+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetRecipeListUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-class GetFoodCategoryImpl:GetFoodCategory {
+class GetRecipeListUseCaseImpl:GetRecipeListUseCase {
     override fun invoke(): List<String> {
         TODO("Not yet implemented")
     }
