--- app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetUserFoodCategoryListImpl.kt
@@ -1,7 +1,0 @@
-package com.angdroid.refrigerator_manament.domain.usecase
-
-class GetUserFoodCategoryListImpl : GetUserFoodCategoryList {
-    override fun invoke(): List<String> {
-        TODO("Not yet implemented")
-    }
-}
