+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetUserFoodCategoryListUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-class GetRecipeDetailImpl : GetRecipeDetail {
+class GetUserFoodCategoryListUseCaseImpl : GetUserFoodCategoryListUseCase {
     override fun invoke(): List<String> {
         TODO("Not yet implemented")
     }
