+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetUserFoodDetailUseCase.kt
@@ -1,0 +1,5 @@
+package com.angdroid.refrigerator_manament.domain.usecase
+
+interface GetUserFoodDetailUseCase { //음식리스트에서 음식의 갯수, 유통기한을 가져올 함수 입니다.
+    operator fun invoke(): Unit
+}
