+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetUserProfileUseCase.kt
@@ -1,0 +1,5 @@
+package com.angdroid.refrigerator_manament.domain.usecase
+
+interface GetUserProfileUseCase { //UserData를 가져올 함수 이후 name:String, Image : String형태가 될 것
+    operator fun invoke() : Unit
+}
