+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/GetUserProfileUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.angdroid.refrigerator_manament.domain.usecase
 
-class GetUserFoodDetailImpl : GetUserFoodDetail {
+class GetUserProfileUseCaseImpl : GetUserProfileUseCase {
     override fun invoke() {
         TODO("Not yet implemented")
     }
