+++ app/src/main/java/com/angdroid/refrigerator_manament/domain/usecase/SetUserFoodDetailUseCaseImpl.kt
@@ -1,0 +1,7 @@
+package com.angdroid.refrigerator_manament.domain.usecase
+
+class SetUserFoodDetailUseCaseImpl : SetUserFoodDetailUseCase {
+    override fun invoke() {
+        TODO("Not yet implemented")
+    }
+}
